cached reused results
it was pissing me off that it took a while to finish the calculation
for the input from the file. after examining, I figured is because the
same line can intersect with multiple other lines. hence, instead of
recomputing the trace like a neanderthal everytime it shows up, just
cache the damn thing.

it now runs as fact as expected.