created instruction wrapper
advent of code was pissing me off with that many requirements for the
intcode machine, that I just decided to take it seriously and make a
fully functional wrapper for machine instructions that come in the form
of integers. this one is neat because it only allocates a small buffer
upon creation but then after that updates itself without having to
allocate memory.