finished day 5!
finally, had a little trouble making sure that both versions of the code
co-existed together. it gave a lot of insight as to what tricks are
available so as to avoid code duplication in rust.

so far, trying to avoid code duplication in rust is a pain in the ass.
9 out 10 times you have to write more code in order to avoid duplicating
code which kinda of defeats the purpose of trying to avoid code
duplication in the first place.

the thing that had me stuck the most, was how to use hook the intcode
compiler with a different set of instructions. as you may have noticed,
there is an instruction set for part-one, and an extended version of the
same instruction set is presented in part-two. trying to make the
compiler instruction set independent was probably the most painful thing
I've done in a while.

I managed to solve the problem by having the instruction sets be part of
a variant, I think this is a good enough solution, but can't help but
feel that there must be something better out there.