fixed day 05 code
finally figured out a way to properly use the type system to make the
code much more cleaner, I am in-love with the end result! it's seems
complicated at a first glance, but I believe that this is because I have
implemented everything in a single file. Once everything gets refactored
into separate modules the part that will be left in this file will look
much simpler.

my mistake was to think that the `Command` type should handle the
responsibility to calculate the stride of the given instruction set.
which made the previous version of the code extremely ugly as I had to
manage all cases from a single function.

what I am doing now, is delegating the responsibility of calculating the
stride to the instruction set itself, which makes perfect sense! this
allows new instruction sets to be added much more easily to the code
while having the `impl Command` untouched.