refactored and clean of day 5
I keep coming back to this, but I just can't conceive that something can
be written in an ugly way in this language (lol). so I keep coming back
to this particular excercise cause so far it's the ugliest one I've
written.

right now its in a much better state, and I'm actually pretty satisfied
with how it has changed. I will move some stuff outside of the file, so
that it looks even cleaner.