finished refactoring of intcode module
day 5, is finally the way I had envisioned, and I'm extremely satisfied
with the way things turned out. parameter processing is solely the
responsibility of the `intcode` module which is nice cause there will
probably be more parameter types in future challenges.

the only thing that the user handles is everything related to the
machine code instructions.